kdc: Use correct key for FAST encrypted challenge
According to RFC6113:

The challenge key used by the client to send a message to the KDC is
KRB-FX-CF2(armor_key, long_term_key, "clientchallengearmor",
"challengelongterm"). The challenge key used by the KDC encrypting to
the client is KRB-FX-CF2(armor_key, long_term_key, "kdcchallengearmor",
"challengelongterm").

Previously, the KDC was encrypting the message using the same key that
it used to decrypt the message from the client.